Added entrySchmea and added journalId to userschema
Instead of including journal entries in the User object. Split
journal entries into a seperate collection with a unique journal
id for each user. Added journalid to userschema so that it will be easy to identify. This allow for easy modelling of the data, and a separate api resource for the journal entries.